// web/src/shared/api/generated/model/requestTrainingExportRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { TrainingExportSource } from './trainingExportSource';

export interface RequestTrainingExportRequest {
  source: TrainingExportSource;
  from: string;
  to: string;
  /** @nullable */
  subjectId: string | null;
}